refactor: fold cancel tests into test_01_hello_app
The cancel_app was nearly identical to 01-hello-app, and cancellation
is always enabled, so a separate app isn't needed.
