+++ pkg-py/tests/playwright/test_01_hello_app.py
@@ -197,3 +197,55 @@ class Test01HelloApp:
         expect(sql_code).to_contain_text(
             re.compile(r"WHERE.*sex.*=.*['\"]?male['\"]?", re.IGNORECASE), timeout=60000
         )
+
+    # ==================== Stream Cancellation Tests ====================
+
+    def test_stop_button_appears_during_streaming(self) -> None:
+        """CANCEL-01: The stop button should be visible while the LLM is streaming."""
+        self.chat.set_user_input(
+            "Write a very long and detailed paragraph about every column "
+            "in this dataset. Be extremely thorough and verbose."
+        )
+        self.chat.send_user_input(method="click")
+
+        stop_btn = self.page.locator(".shiny-chat-btn-cancel")
+        expect(stop_btn).to_be_visible(timeout=30000)
+
+    def test_cancel_stops_response(self) -> None:
+        """CANCEL-02: Clicking the stop button should cancel the stream."""
+        self.chat.set_user_input(
+            "Write a very long and detailed paragraph about every column "
+            "in this dataset. Be extremely thorough and verbose."
+        )
+        self.chat.send_user_input(method="click")
+
+        stop_btn = self.page.locator(".shiny-chat-btn-cancel")
+        expect(stop_btn).to_be_visible(timeout=30000)
+        stop_btn.click()
+
+        cancelled = self.page.locator(".shiny-chat-message-cancelled")
+        expect(cancelled).to_be_visible(timeout=15000)
+        expect(cancelled).to_have_text("Response cancelled")
+
+    def test_can_send_after_cancel(self) -> None:
+        """CANCEL-03: After cancelling, the user should be able to send another message."""
+        self.chat.set_user_input(
+            "Write a very long and detailed paragraph about every column "
+            "in this dataset. Be extremely thorough and verbose."
+        )
+        self.chat.send_user_input(method="click")
+
+        stop_btn = self.page.locator(".shiny-chat-btn-cancel")
+        expect(stop_btn).to_be_visible(timeout=30000)
+        stop_btn.click()
+
+        cancelled = self.page.locator(".shiny-chat-message-cancelled")
+        expect(cancelled).to_be_visible(timeout=15000)
+
+        self.chat.set_user_input("How many rows are in the dataset?")
+        self.chat.send_user_input(method="click")
+
+        send_btn = self.page.locator(
+            ".shiny-chat-btn-send:not(.shiny-chat-btn-cancel)"
+        )
+        expect(send_btn).to_be_visible(timeout=60000)
